Split the massive Handlers/Events modules into a few more granular pieces
