+++ surveyor-core/src/Surveyor/Core/Loader.hs
@@ -43,7 +43,7 @@ import qualified What4.Expr.Builder as WEB
 import qualified Surveyor.Core.Architecture as A
 import           Surveyor.Core.BinaryAnalysisResult
 import qualified Surveyor.Core.Chan as C
-import           Surveyor.Core.Events ( Events(..) )
+import qualified Surveyor.Core.Events as SCE
 import qualified Surveyor.Core.Loader.RenovateAnalysis as RA
 import qualified Surveyor.Core.Loader.PPCConfig as LP
 
@@ -64,14 +64,14 @@ cancelLoader al = do
 
 -- | Try to load the given file, attempting to determine its type based on the
 -- filename
-asynchronouslyLoad :: NG.NonceGenerator IO s -> C.Chan (Events s st) -> FilePath -> IO AsyncLoader
+asynchronouslyLoad :: NG.NonceGenerator IO s -> C.Chan (SCE.Events s st) -> FilePath -> IO AsyncLoader
 asynchronouslyLoad ng customEventChan path
   | takeExtension path == ".bc" = asynchronouslyLoadLLVM ng customEventChan path
   | takeExtension path == ".jar" = asynchronouslyLoadJAR ng customEventChan path
   | otherwise = asynchronouslyLoadElf ng customEventChan path
 
 -- | Load a JAR file
-asynchronouslyLoadJAR :: NG.NonceGenerator IO s -> C.Chan (Events s st) -> FilePath -> IO AsyncLoader
+asynchronouslyLoadJAR :: NG.NonceGenerator IO s -> C.Chan (SCE.Events s st) -> FilePath -> IO AsyncLoader
 asynchronouslyLoadJAR ng customEventChan jarPath = do
   nonce <- NG.freshNonce ng
   jr0 <- A.mkInitialJVMResult nonce
@@ -83,12 +83,12 @@ asynchronouslyLoadJAR ng customEventChan jarPath = do
       -- The loader chunks the class list so that it can report incremental
       -- progress to the user
       lastRes <- F.foldlM (addJARChunk jr) jr0 chunks
-      C.writeChan customEventChan (AnalysisFinished (A.SomeResult lastRes) [])
+      SCE.emitEvent customEventChan (SCE.AnalysisFinished (A.SomeResult lastRes) [])
     C.putMVar mv worker
     eres <- A.waitCatch worker
     case eres of
       Right () -> return ()
-      Left exn -> C.writeChan customEventChan (AnalysisFailure exn)
+      Left exn -> SCE.emitEvent customEventChan (SCE.AnalysisFailure exn)
   worker <- C.takeMVar mv
   return AsyncLoader { errorCatcher = errThread
                      , workerThread = worker
@@ -98,10 +98,10 @@ asynchronouslyLoadJAR ng customEventChan jarPath = do
       let readClass className = J.loadClassFromJar className jr
       classes <- catMaybes <$> mapM readClass classNames
       res' <- A.extendJVMResult res classes
-      C.writeChan customEventChan (AnalysisProgress (A.SomeResult res'))
+      SCE.emitEvent customEventChan (SCE.AnalysisProgress (A.SomeResult res'))
       return res'
 
-asynchronouslyLoadLLVM :: NG.NonceGenerator IO s -> C.Chan (Events s st) -> FilePath -> IO AsyncLoader
+asynchronouslyLoadLLVM :: NG.NonceGenerator IO s -> C.Chan (SCE.Events s st) -> FilePath -> IO AsyncLoader
 asynchronouslyLoadLLVM ng customEventChan bcPath = do
   mv <- C.newEmptyMVar
   errThread <- A.async $ do
@@ -110,17 +110,17 @@ asynchronouslyLoadLLVM ng customEventChan bcPath = do
       em <- LL.parseBitCode bs
       case em of
         Left err ->
-          C.writeChan customEventChan (ErrorLoadingLLVM (LL.formatError err))
+          SCE.emitEvent customEventChan (SCE.ErrorLoadingLLVM (LL.formatError err))
         Right m -> do
           nonce <- NG.freshNonce ng
           hdlAlloc <- CFH.newHandleAllocator
           llvmRes <- A.mkLLVMResult ng nonce hdlAlloc m
-          C.writeChan customEventChan (AnalysisFinished llvmRes [])
+          SCE.emitEvent customEventChan (SCE.AnalysisFinished llvmRes [])
     C.putMVar mv worker
     eres <- A.waitCatch worker
     case eres of
       Right () -> return ()
-      Left exn -> C.writeChan customEventChan (AnalysisFailure exn)
+      Left exn -> SCE.emitEvent customEventChan (SCE.AnalysisFailure exn)
   worker <- C.takeMVar mv
   return AsyncLoader { errorCatcher = errThread
                      , workerThread = worker
@@ -130,7 +130,7 @@ asynchronouslyLoadLLVM ng customEventChan bcPath = do
 --
 -- The thread sends the result of loading to the main thread through the
 -- provided event channel
-asynchronouslyLoadElf :: NG.NonceGenerator IO s -> C.Chan (Events s st) -> FilePath -> IO AsyncLoader
+asynchronouslyLoadElf :: NG.NonceGenerator IO s -> C.Chan (SCE.Events s st) -> FilePath -> IO AsyncLoader
 asynchronouslyLoadElf ng customEventChan exePath = do
   mv <- C.newEmptyMVar
   thread <- A.async $ do
@@ -140,16 +140,16 @@ asynchronouslyLoadElf ng customEventChan exePath = do
       bs <- BS.readFile exePath
       case E.parseElf bs of
         E.ElfHeaderError off msg ->
-          C.writeChan customEventChan (ErrorLoadingELFHeader off msg)
+          SCE.emitEvent customEventChan (SCE.ErrorLoadingELFHeader off msg)
         E.Elf32Res [] e32 -> loadElf ng customEventChan (E.Elf32 e32)
         E.Elf64Res [] e64 -> loadElf ng customEventChan (E.Elf64 e64)
-        E.Elf32Res errs _ -> C.writeChan customEventChan (ErrorLoadingELF errs)
-        E.Elf64Res errs _ -> C.writeChan customEventChan (ErrorLoadingELF errs)
+        E.Elf32Res errs _ -> SCE.emitEvent customEventChan (SCE.ErrorLoadingELF errs)
+        E.Elf64Res errs _ -> SCE.emitEvent customEventChan (SCE.ErrorLoadingELF errs)
     C.putMVar mv worker
     eres <- A.waitCatch worker
     case eres of
       Right () -> return ()
-      Left exn -> C.writeChan customEventChan (AnalysisFailure exn)
+      Left exn -> SCE.emitEvent customEventChan (SCE.AnalysisFailure exn)
   worker <- C.takeMVar mv
   return AsyncLoader { errorCatcher = thread
                      , workerThread = worker
@@ -159,7 +159,7 @@ asynchronouslyLoadElf ng customEventChan exePath = do
 -- across all of the streamed results for the same executable.  We need to
 -- generate the nonce under 'withElfConfig' so that we can capture the
 -- appropriate value of w.
-loadElf :: NG.NonceGenerator IO s -> C.Chan (Events s st) -> E.SomeElf E.Elf -> IO ()
+loadElf :: NG.NonceGenerator IO s -> C.Chan (SCE.Events s st) -> E.SomeElf E.Elf -> IO ()
 loadElf ng customEventChan someElf = do
   -- FIXME: Make the float mode configurable
   sym <- SB.newSimpleBackend WEB.FloatRealRepr ng
@@ -177,7 +177,7 @@ loadElf ng customEventChan someElf = do
                                            }
             let sr = A.mkX86Result res
             void $ X.evaluate (DS.force sr)
-            C.writeChan customEventChan (AnalysisProgress sr)
+            SCE.emitEvent customEventChan (SCE.AnalysisProgress sr)
   let x86cfg = x86cfg0 { R.rcFunctionCallback = Just (10, x86callback) }
   ppc32cfg <- LP.ppcConfig sym MBL.Elf32Repr customEventChan ng PPC32.allDefinedFunctions PPC32.allSemantics PPC.config32 A.mkPPC32Result
   ppc64cfg <- LP.ppcConfig sym MBL.Elf64Repr customEventChan ng PPC64.allDefinedFunctions PPC64.allSemantics PPC.config64 A.mkPPC64Result
@@ -188,4 +188,4 @@ loadElf ng customEventChan someElf = do
   R.withElfConfig someElf rcfgs $ \rc elf loadedBinary -> do
     hdlAlloc <- CFH.newHandleAllocator
     (res, diags) <- R.analyzeElf rc hdlAlloc elf loadedBinary
-    C.writeChan customEventChan (AnalysisFinished res diags)
+    SCE.emitEvent customEventChan (SCE.AnalysisFinished res diags)
